+++ terraform/tig_ecr.tf
@@ -1,0 +1,44 @@
+data "aws_ecr_authorization_token" "token" {}
+
+locals {
+  lambda_container_image_uri_split = split("/", var.lambda_container_image_uri)
+  ecr_image_name_and_tag           = split(":", element(local.lambda_container_image_uri_split, length(local.lambda_container_image_uri_split) - 1))
+  ecr_image_name                   = "${local.environment}-${element(local.ecr_image_name_and_tag, 0)}"
+  ecr_image_tag                    = element(local.ecr_image_name_and_tag, 1)
+}
+
+resource aws_ecr_repository "lambda-image-repo" {
+  name = local.ecr_image_name
+  tags = var.tags
+}
+
+
+resource null_resource ecr_login {
+  triggers = {
+    image_uri = var.lambda_container_image_uri
+  }
+
+  provisioner "local-exec" {
+    interpreter = ["/bin/bash", "-e", "-c"]
+    command = <<EOF
+      docker login ${data.aws_ecr_authorization_token.token.proxy_endpoint} -u AWS -p ${data.aws_ecr_authorization_token.token.password}
+      EOF
+  }
+}
+
+
+resource null_resource upload_ecr_image {
+  depends_on = [null_resource.ecr_login]
+  triggers = {
+    image_uri = var.lambda_container_image_uri
+  }
+
+  provisioner "local-exec" {
+    interpreter = ["/bin/bash", "-e", "-c"]
+    command = <<EOF
+      docker pull ${var.lambda_container_image_uri}
+      docker tag ${var.lambda_container_image_uri} ${aws_ecr_repository.lambda-image-repo.repository_url}:${local.ecr_image_tag}
+      docker push ${aws_ecr_repository.lambda-image-repo.repository_url}:${local.ecr_image_tag}
+      EOF
+  }
+}